# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Lynx Analytics — Plataforma para micro-SaaS brasileiros
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /projects/{projectId}/logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List application logs
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warn, error]
        - name: service
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Logs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogListResponse"
    post:
      operationId: ingestLog
      tags: [logs]
      summary: Ingest log entries
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestLogRequest"
      responses:
        "201":
          description: Logs ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestLogResponse"
  /projects/{projectId}/requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List HTTP requests
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
        - name: statusCode
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Requests list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
    post:
      operationId: ingestRequest
      tags: [requests]
      summary: Ingest HTTP request data
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequestRequest"
      responses:
        "201":
          description: Request ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestRecord"
...
  /projects/{projectId}/funnels:
    get:
      operationId: listFunnels
      tags: [funnels]
      summary: List funnels for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Funnels list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Funnel"
    post:
      operationId: createFunnel
      tags: [funnels]
      summary: Create a funnel
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunnelRequest"
      responses:
        "201":
          description: Created funnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funnel"
  /projects/{projectId}/funnels/{funnelId}/results:
    get:
      operationId: getFunnelResults
      tags: [funnels]
      summary: Get funnel results and conversion rates
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: funnelId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Funnel results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelResults"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        apiKey:
          type: string
        description:
          type: string
        website:
          type: string
        abacatePayConnected:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        eventCount:
          type: integer
        sessionCount:
          type: integer
        mrr:
          type: number
        customerCount:
          type: integer
      required: [id, name, slug, apiKey, createdAt, updatedAt, abacatePayConnected]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
      required: [name]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        service:
          type: string
        timestamp:
          type: string
          format: date-time
        meta:
          type: object
          additionalProperties: true
        traceId:
          type: string
        spanId:
          type: string
      required: [id, projectId, level, message, timestamp]
    IngestLogRequest:
      type: object
      properties:
        entries:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
                enum: [debug, info, warn, error]
              message:
                type: string
              service:
                type: string
              timestamp:
                type: string
                format: date-time
              meta:
                type: object
                additionalProperties: true
              traceId:
                type: string
            required: [level, message]
      required: [entries]
    IngestLogResponse:
      type: object
      properties:
        count:
          type: integer
      required: [count]
    LogListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [entries, total, limit, offset]
    RequestRecord:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        method:
          type: string
        url:
          type: string
        statusCode:
          type: integer
        duration:
          type: number
        requestSize:
          type: integer
        responseSize:
          type: integer
        timestamp:
          type: string
          format: date-time
        ip:
          type: string
        userAgent:
          type: string
        traceId:
          type: string
        error:
          type: string
      required: [id, projectId, method, url, statusCode, duration, timestamp]
    IngestRequestRequest:
      type: object
      properties:
        method:
          type: string
        url:
          type: string
        statusCode:
          type: integer
        duration:
          type: number
        requestSize:
          type: integer
        responseSize:
          type: integer
        timestamp:
          type: string
          format: date-time
        ip:
          type: string
        userAgent:
          type: string
        traceId:
          type: string
        error:
          type: string
      required: [method, url, statusCode, duration]
    RequestListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/RequestRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [requests, total, limit, offset]
...
    Funnel:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStep"
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, name, steps, createdAt]
    FunnelStep:
      type: object
      properties:
        order:
          type: integer
        eventName:
          type: string
        label:
          type: string
      required: [order, eventName, label]
    CreateFunnelRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
            properties:
              order:
                type: integer
              eventName:
                type: string
              label:
                type: string
            required: [order, eventName, label]
      required: [name, steps]
    FunnelResults:
      type: object
      properties:
        funnelId:
          type: string
        funnelName:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStepResult"
        overallConversionRate:
          type: number
        period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
          required: [from, to]
      required: [funnelId, funnelName, steps, overallConversionRate, period]
    FunnelStepResult:
      type: object
      properties:
        order:
          type: integer
        label:
          type: string
        eventName:
          type: string
        users:
          type: integer
        conversionRate:
          type: number
        dropoffRate:
          type: number
      required: [order, label, eventName, users, conversionRate, dropoffRate]
...